fix: ensure a stable sort across pages
If doing pagination, the order is important since you're relying on it being stable across pages to actually walk through the intended dataset.

If the ORDER BY is not specified, the actual order is undefined, meaning you can get pages skipped and duplicated. It will mostly work because the same order is kept, but sometimes it might fail (as in my case, on MariaDB 10). The issue is, it didn't always fail, only sometimes (on some pages), being totally unpredictable and requiring a large amount of debugging.